+++ app/src/main/java/cn/etsoft/smarthome/Domain/WareSetBox.java
@@ -9,7 +9,7 @@ public class WareSetBox implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 	private WareDev dev;
-	private byte bOnOff;
+	private int bOnOff;
 
 	public static long getSerialVersionUID() {
 		return serialVersionUID;
@@ -23,11 +23,11 @@ public class WareSetBox implements Serializable {
 		this.dev = dev;
 	}
 
-	public byte getbOnOff() {
+	public int getbOnOff() {
 		return bOnOff;
 	}
 
-	public void setbOnOff(byte bOnOff) {
+	public void setbOnOff(int bOnOff) {
 		this.bOnOff = bOnOff;
 		dev.setbOnOff(bOnOff);
 	}
